change type of port to unit16
In modern computers, the port numbers are between 1 and 65535, so unit16 is
sufficient.

+++ web/web.go
@@ -24,7 +24,7 @@ func mainHandler(c http.ResponseWriter, req *http.Request) {
 	mainTempl.Execute(c, p)
 }
 
-func Start(server *raft.Server, port int) {
+func Start(server *raft.Server, port uint16) {
 	mainTempl = template.Must(template.New("index.html").Parse(index_html))
 	s = server
